refactor: remove PSection, add PContainer
- remove PLayout to PRow
