+++ src/core/components/Structure/PContainer.vue
@@ -1,11 +1,15 @@
 <script>
-/*short code PSc*/
+/*short code PCt*/
 
 export default {
   functional: true,
   render(h, context) {
 
-    return h('section', {
+    return h('v-container', {
+        props: {
+            fluid: context.props.cnf.a,
+            tag: context.props.cnf.b
+        },
         //class: context.props.cnf.b||'',
         attrs: context.data.attrs,
         class: context.data.class
