+++ src/core/components/Structure/index.js
@@ -3,7 +3,7 @@ export {default as PAl}  from  './PAlert.vue';
 export {default as PAv}  from  './PAvatar.vue';
 export {default as PCd}  from  './PCard.vue';
 export {default as PCs}  from  './PCarusel.vue';
-export {default as PSc}  from  './PSection.vue';
+export {default as PCt}  from  './PContainer.vue';
 export {default as PDl}  from  './PDialog.vue';
 export {default as PDm}  from  './PDialogMenu';
 export {default as PEx}  from  './PExpansion.vue';
@@ -26,7 +26,7 @@ export {default as PTm}  from  './PTimeline.vue';
 export {default as PTr}  from  './PToolbar.vue';
 export {default as PYv}  from  './PYoutubeVideo.vue';
 export {default as PAd}  from  './PAdsense.vue';
-export {default as PPn}  from  './PSiteoPlugin.vue'; 
+export {default as PPn}  from  './PSiteoPlugin.vue';
 export const RSp = 'VSpacer';
 export const RDh = 'VDivider'
 export const RDv = {
