+++ src/core/components/index.js
@@ -1,7 +1,7 @@
 // Global Elements
 export {default as PBt} from './Structure/PButton.js';
 export {default as SiteoPlugin} from './SiteoPlugin.vue';
-export {default as PLt}  from  './Structure/PLayout.vue';
+export {default as PLt}  from  './Structure/PRow.vue';
 export {default as AppPulseLoader} from './Elements/AppPulseLoader.vue';
 export {default as AppErrorResponse} from './Elements/AppErrorResponse/index.vue';
 export {default as AppMessagesBlock} from './Elements/AppMessagesBlock.js';
